+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/CircuitBreakerService.java
@@ -1,9 +1,10 @@
 package com.mageddo.dnsproxyserver.solver.remote;
 
+import java.net.InetSocketAddress;
 import java.util.function.Supplier;
 
 public interface CircuitBreakerService {
 
-  Result handle(Request req, Supplier<Result> sup);
+  Result safeHandle(final InetSocketAddress resolverAddress, Supplier<Result> sup);
 
 }