+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/application/CircuitBreakerFailSafeService.java
@@ -2,7 +2,6 @@ package com.mageddo.dnsproxyserver.solver.remote.application;
 
 import com.mageddo.commons.circuitbreaker.CircuitCheckException;
 import com.mageddo.dnsproxyserver.solver.remote.CircuitBreakerService;
-import com.mageddo.dnsproxyserver.solver.remote.Request;
 import com.mageddo.dnsproxyserver.solver.remote.Result;
 import dev.failsafe.CircuitBreakerOpenException;
 import dev.failsafe.Failsafe;
@@ -12,6 +11,7 @@ import org.apache.commons.lang3.ClassUtils;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
+import java.net.InetSocketAddress;
 import java.util.function.Supplier;
 
 @Slf4j
@@ -24,21 +24,29 @@ public class CircuitBreakerFailSafeService implements CircuitBreakerService {
   private String status;
 
   @Override
-  public Result handle(Request req, Supplier<Result> sup) {
-    final var circuitBreaker = this.circuitBreakerFactory.createCircuitBreakerFor(req.getResolverAddress());
+  public Result safeHandle(InetSocketAddress resolverAddress, Supplier<Result> sup) {
     try {
-      return Failsafe
-        .with(circuitBreaker)
-        .get((ctx) -> sup.get());
+      return this.handle(resolverAddress, sup);
     } catch (CircuitCheckException | CircuitBreakerOpenException e) {
       final var clazz = ClassUtils.getSimpleName(e);
-      log.debug("status=circuitEvent, server={}, type={}", req.getResolverAddress(), clazz);
-      this.status = String.format("%s for %s", clazz, req.getResolverAddress());
+      log.debug("status=circuitEvent, server={}, type={}", resolverAddress, clazz);
+      this.status = String.format("%s for %s", clazz, resolverAddress);
       return Result.empty();
     }
   }
 
+  private Result handle(InetSocketAddress resolverAddress, Supplier<Result> sup) {
+    final var circuitBreaker = this.circuitBreakerFactory.createCircuitBreakerFor(resolverAddress);
+    return Failsafe
+      .with(circuitBreaker)
+      .get((ctx) -> sup.get());
+  }
+
   public String getStatus() {
     return this.status;
+  }
+
+  public void resetCircuitBreakerFactory() {
+    this.circuitBreakerFactory.reset();
   }
 }