+++ src/test/java/testing/Events.java
@@ -11,9 +11,11 @@ import lombok.extern.slf4j.Slf4j;
 public class Events implements EventHandler<Context> {
   @Override
   public void afterSetup(Context component) {
+
     final var config = Configs.getInstance();
     log.info("status=startingDPS, port={}", config.getWebServerPort());
     component.start();
+
     RestAssured.port = config.getWebServerPort();
     RestAssured.config = RestAssured
         .config()