+++ src/test/java/testing/templates/CircuitBreakerConfigTemplates.java
@@ -3,7 +3,7 @@ package testing.templates;
 import com.mageddo.dnsproxyserver.config.CircuitBreaker;
 import com.mageddo.dnsproxyserver.config.mapper.ConfigMapper;
 
-public class CircuitBreakerTemplates {
+public class CircuitBreakerConfigTemplates {
   public static CircuitBreaker buildDefault(){
     return ConfigMapper.defaultCircuitBreaker();
   }