+++ src/test/java/testing/templates/solver/remote/FailSafeCircuitBreakerTemplates.java
@@ -1,0 +1,15 @@
+package testing.templates.solver.remote;
+
+import com.mageddo.commons.circuitbreaker.CircuitCheckException;
+import com.mageddo.dnsproxyserver.solver.remote.Result;
+import dev.failsafe.CircuitBreaker;
+
+public class FailSafeCircuitBreakerTemplates {
+  public static CircuitBreaker<Result> buildDefault() {
+    return CircuitBreaker.<Result>builder()
+      .handle(CircuitCheckException.class)
+      .withFailureThreshold(1)
+      .withSuccessThreshold(2)
+      .build();
+  }
+}